Reduce use of `-XCPP` for conditional compilation (#168)
