+++ nix/pkgs.nix
@@ -1,9 +1,13 @@
-{ compiler, enableDhall, enableSwagger }:
+{ compiler
+, enableDhall
+, enableSwagger
+, swaggerWrapperFormat
+}:
 
 import ./nixpkgs.nix {
   overlays = [
     (import ./overlays/haskell-packages.nix {
-      inherit compiler enableDhall enableSwagger;
+      inherit compiler enableDhall enableSwagger swaggerWrapperFormat;
     })
   ];
 }
