+++ proto3-suite.cabal
@@ -46,10 +46,13 @@ library
 
   if flag(swagger)
     exposed-modules:   Proto3.Suite.DotProto.Generate.Swagger
+                       Proto3.Suite.DotProto.Generate.Swagger.Wrappers
     build-depends:     swagger2 >=2.1.6 && <2.7
     cpp-options:       -DSWAGGER
     if flag(swagger-wrapper-format)
-      cpp-options:       -DSWAGGER_WRAPPER_FORMAT
+      hs-source-dirs:  src/swagger-wrapper-format
+    else
+      hs-source-dirs:  src/no-swagger-wrapper-format
 
   exposed-modules:     Proto3.Suite
                        Proto3.Suite.Class
