+++ src/Proto3/Suite/DotProto/Generate/Swagger.hs-boot
@@ -1,0 +1,6 @@
+module Proto3.Suite.DotProto.Generate.Swagger
+  ( OverrideToSchema (..)
+  )
+where
+
+newtype OverrideToSchema a = OverrideToSchema { unOverride :: a }
