+++ src/no-swagger-wrapper-format/Proto3/Suite/DotProto/Generate/Swagger/Wrappers.hs
@@ -1,0 +1,1 @@
+module Proto3.Suite.DotProto.Generate.Swagger.Wrappers where
